requirements.txt
python-telegram-bot==20.0a2
python-dotenv 
requests
dnspython 
pymongo 